django_api/requirements.txt
django==3.2
flake8==4.0
python-dotenv==0.21.0
psycopg2-binary==2.9.6
django-split-settings==1.2.0
djangorestframework==3.14.0
gunicorn==20.0.4
django-cors-headers==2.2.0
pydantic==1.10.9
elasticsearch==8.8.0